HIPCMS-472 cleanup
Corrected Status codes,
Added Permission for invitation

+++ src/Api/Controllers/UsersController.cs
@@ -1,14 +1,13 @@
 ﻿using Api.Utility;
-using Api.Models.Entity;
 using Microsoft.AspNetCore.Mvc;
 using Api.Managers;
 using Microsoft.Extensions.Logging;
-using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Http;
 using System.IO;
 using Api.Data;
 using Api.Models;
 using Api.Models.User;
+using Api.Permission;
 
 namespace Api.Controllers
 {
@@ -16,20 +15,29 @@ namespace Api.Controllers
     {
         private UserManager userManager;
         private EmailSender emailSender;
+        private UserPermissions userPermissions;
 
         public UsersController(CmsDbContext dbContext, EmailSender emailSender, ILoggerFactory _logger) : base(dbContext, _logger)
         {
             userManager = new UserManager(dbContext);
+            userPermissions = new UserPermissions(dbContext);
             this.emailSender = emailSender;
         }
 
+        #region invite
+
         // POST api/users/invite
         [HttpPost("Invite")]
-        [Authorize(Roles = Role.Supervisor + "," + Role.Administrator)]
-        [ProducesResponseType(typeof(void), 200)]
+        [ProducesResponseType(typeof(void), 202)]
         [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 403)]
+        [ProducesResponseType(typeof(void), 409)]
+        [ProducesResponseType(typeof(void), 503)]
         public IActionResult Post(InviteFormModel model)
         {
+            if (!userPermissions.IsAllowedToInvite(User.Identity.GetUserId()))
+                return Forbidden();
+
             if (ModelState.IsValid)
             {
                 int failCount = 0;
@@ -60,11 +68,12 @@ namespace Api.Controllers
             return BadRequest(ModelState);
         }
 
+        #endregion
+
         #region GET user
         // GET api/users
         [HttpGet]
         [ProducesResponseType(typeof(PagedResult<UserResult>), 200)]
-        [ProducesResponseType(typeof(void), 404)]
         public IActionResult Get(string query, string role, int page = 1)
         {
             var users = userManager.GetAllUsers(query, role, page, Constants.PageSize);
@@ -110,6 +119,7 @@ namespace Api.Controllers
         [HttpPut("Current")]
         [ProducesResponseType(typeof(void), 200)]
         [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 404)]
         public IActionResult Put(UserFormModel model)
         {
             return PutUser(User.Identity.GetUserId(), model);
@@ -117,11 +127,15 @@ namespace Api.Controllers
 
         // PUT api/users/5
         [HttpPut("{id}")]
-        [Authorize(Roles = Role.Administrator)]
         [ProducesResponseType(typeof(void), 200)]
         [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 403)]
+        [ProducesResponseType(typeof(void), 404)]
         public IActionResult Put(int id, AdminUserFormModel model)
         {
+            if (!userPermissions.IsAllowedToAdminister(User.Identity.GetUserId()))
+                return Forbidden();
+
             return PutUser(id, model);
         }
 
@@ -140,6 +154,7 @@ namespace Api.Controllers
                         _logger.LogInformation(5, "User with ID: " + id + " updated.");
                         return Ok();
                     }
+                    return NotFound();
                 }
             }
 
@@ -153,7 +168,7 @@ namespace Api.Controllers
         // GET api/users/{userId}/picture/
         [HttpGet("{userId}/picture/")]
         [ProducesResponseType(typeof(string), 200)]
-        [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 404)]
         public IActionResult GetPictureById(int userId)
         {
             return GetPicture(userId);
@@ -162,7 +177,7 @@ namespace Api.Controllers
         // GET api/users/current/picture/
         [HttpGet("Current/picture/")]
         [ProducesResponseType(typeof(string), 200)]
-        [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 404)]
         public IActionResult GetPictureForCurrentUser()
         {
             return GetPicture(User.Identity.GetUserId());
@@ -179,7 +194,7 @@ namespace Api.Controllers
 
                 return Ok(ToBase64String(path));
             }
-            return BadRequest();
+            return NotFound();
         }
 
         #endregion
@@ -189,11 +204,13 @@ namespace Api.Controllers
 
         // Post api/users/{id}/picture/
         [HttpPost("{id}/picture/")]
-        [Authorize(Roles = Role.Administrator)]
         [ProducesResponseType(typeof(void), 200)]
         [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 403)]
         public IActionResult PutPicture(int id, IFormFile file)
         {
+            if (!userPermissions.IsAllowedToAdminister(User.Identity.GetUserId()))
+                return Forbidden();
             return PutUserPicture(id, file);
         }
 
@@ -246,11 +263,13 @@ namespace Api.Controllers
         #region DELETE picture
 
         [HttpDelete("{id}/picture/")]
-        [Authorize(Roles = Role.Administrator)]
         [ProducesResponseType(typeof(void), 200)]
         [ProducesResponseType(typeof(void), 400)]
+        [ProducesResponseType(typeof(void), 403)]
         public IActionResult Delete(int id)
         {
+            if (!userPermissions.IsAllowedToAdminister(User.Identity.GetUserId()))
+                return Forbidden();
             return DeletePicture(id);
         }
